Implement functionality for letting a user pick their opponent from a list of their friends, when creating a dare
The dropdown for selecting a dare opponent no longer contains hardcoded values.
Now the dropdown is filled with the users friends.
Also some minor cleanup in the form widget tree.

Changed getFriendsList in the userService class to return a list instead of a map.
The return data is now more logically structured and easier to work with.
Also updated the associated test to reflect this change.