Implement showing the logged in user's dares on the home screen
The home screen now shows the logged in users dares, fetched from
the server, instead of hardcoded values.
Changed the return of the getDares method in the DareLogic class,
it now returns a list instead of a map.
Updated associated tests to reflect changes in the DareLogic class
The widget tree now provides a DareLogic object instead of a DareService object,
this is not a final change.