Fix the user being signed in and out in the gui after registration fails
Fixed a bug where the gui would update to a signed in state and then
back to the starting screen. This happened when a user registration
failed as a result of the dare server was down.
Now the user stays on the registration page and receives an error
message instead.